Compressions
A lot about how I/O can be reduced with Compression, can be done on Client or Serverside.

Some formats benefit more from Compression (as is with Text), compared to say JPEGs.

Compression is most worth it in a lot more reads than writes.

Note: Do not compress on BOTH client and Server.

It's better to compress in I/O bound situations, than CPU-bound, because you have to Uncompress/Compress and Load/unload, etc.