Adding Index/Changing Keys, Columns, Tablespaces, Tables, Foreign Keys
A lot of DDL operations.

Operations that mean to state wether something is available in terms of being in place, or a reconstruction of the table is needed.

Optimization tricks in terms of adding indexes after the fact of a Table alotment.

Ability to change names without affecting Table reconstructions, not having to share locks, etc.

In general, interacting with Solid states is not a in place option, where of Virtual interactions can be done in place.